Make gossipsub backward compatibility its own file
